+++ CMakeLists.txt
@@ -16,19 +16,31 @@ add_subdirectory(extern/pybind11)
 # This is completely optional and just adds hints to IDEs - no affect on build at all.
 file(GLOB_RECURSE BOOST_HIST_FILES "extern/histogram/include/*.hpp")
 
+# List the header files for the Python extension
+set(BOOST_HIST_PY_HEADERS
+    include/boost/histogram/python/axis.hpp
+    include/boost/histogram/python/pybind11.hpp
+    include/boost/histogram/python/storage.hpp
+)
+
+# List the source files for the Python extension
+set(BOOST_HIST_PY_SRC
+     src/module.cpp
+
+     src/histogram/axis.cpp
+     src/histogram/histogram.cpp
+     src/histogram/storage.cpp
+    )
+
 # This is the Python module
 pybind11_add_module(boosthistogram SHARED
-    src/getpython.hpp
-    src/axis.hpp
-    src/module.cpp
-    src/axis.cpp
-    src/histogram.cpp
-    src/storage.cpp
+    ${BOOST_HIST_PY_HEADERS}
+    ${BOOST_HIST_PY_SRC}
     ${BOOST_HIST_FILES}
 )
 
-# This makes IDE's like XCode mimic the Boost Histogram structure
-source_group(TREE ${CMAKE_CURRENT_SOURCE_DIR}/extern/histogram/include PREFIX "Header Files" FILES ${BOOST_HIST_FILES})
+# Add the include directory for boost/histogram/python
+target_include_directories(boosthistogram PRIVATE include)
 
 # These are the Boost header-only libraries required by Boost::Histogram
 target_include_directories(boosthistogram PUBLIC
@@ -54,6 +66,10 @@ target_include_directories(boosthistogram PUBLIC
     extern/variant/include
     )
 
+# This makes IDE's like XCode mimic the Boost Histogram structure
+source_group(TREE ${CMAKE_CURRENT_SOURCE_DIR}/extern/histogram/include PREFIX "Header Files" FILES ${BOOST_HIST_FILES})
+source_group(TREE ${CMAKE_CURRENT_SOURCE_DIR}/include PREFIX "Header Files" FILES ${BOOST_HIST_PY_HEADERS})
+source_group(TREE ${CMAKE_CURRENT_SOURCE_DIR}/src PREFIX "Source Files" FILES ${BOOST_HIST_PY_SRC})
 
 # Tests (Require pytest to be available)
 include(CTest)
