+++ src/histogram/axis.cpp
@@ -4,9 +4,9 @@
 // (See accompanying file LICENSE_1_0.txt
 // or copy at http://www.boost.org/LICENSE_1_0.txt)
 
-#include "getpython.hpp"
+#include <boost/histogram/python/pybind11.hpp>
 
-#include "axis.hpp"
+#include <boost/histogram/python/axis.hpp>
 
 #include <boost/histogram/axis/ostream.hpp>
 #include <boost/histogram.hpp>
